Origin isolation: add various navigation WPTs
This adds web platform tests covering various navigation scenarios for
origin isolation. The intention is to cover at least all the non-popup
navigation cases in isolated_origin_browsertest.cc. We also want to
approach the same coverage for navigation as we have for two-iframe
scenarios (i.e. the various parent-child1-child2 tests), but we avoid
doubling the size of the test suite by focusing on the more important
or tricky scenarios instead of just copying all of those.

Bug: 1042415
Change-Id: Ifb937f183af2610f348e4fea036d96aa0504ddab