Construct StyleResolver in StyleEngine constructor.
Years ago, the StyleResolver was deleted and re-created for certain
style sheet changes. The current lifetime of the StyleResolver is that it
is created once on first EnsureStyleResolver and only freed when the
StyleEngine is disposed. Instead of checking if it needs to be
constructed every time we access it, create it as part of constructing
the StyleEngine.

We don't create it for documents without a Frame for which we should not
use the StyleResolver.

Change-Id: I1f1d2ef322d2599737f584191550f769dcf6828e